requirements.txt
blueprint
Flask
python-dotenv
